… src/Nethermind/Nethermind.Evm.Benchmark/EvmStackPopPushBenchmarks.cs
… src/Nethermind/Nethermind.Evm/EvmState.cs
… src/Nethermind/Nethermind.Evm/Word.cs