This PR reintroduces `Word` structure that was removed due to an erroneous behavior in https://github.com/NethermindEth/nethermind/pull/2816. The EvmStack is rewritten to use the `Word`. Whether `Uint256` could be used instead, should be discussed before this PR is merged. Possibly, the naming of `ulong` fields should be aligned with `Uint256`.

### Setup 

``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.1379 (1909/November2018Update/19H2)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.103
  [Host]        : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  .NET Core 5.0 : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT

Job=.NET Core 5.0  Runtime=.NET Core 5.0  
```

### Before

|   Method |                    v |      Mean |     Error |    StdDev |    Median | Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|--------- |--------------------- |----------:|----------:|----------:|----------:|------:|------:|------:|----------:|----------:|
|     **Byte** |                    **?** | **11.113 ns** | **0.2380 ns** | **0.6752 ns** | **11.001 ns** |     **-** |     **-** |     **-** |         **-** |     **661 B** |
| PushZero |                    ? |  7.802 ns | 0.1644 ns | 0.3167 ns |  7.706 ns |     - |     - |     - |         - |     573 B |
|  PushOne |                    ? |  8.400 ns | 0.2100 ns | 0.5957 ns |  8.221 ns |     - |     - |     - |         - |     577 B |
|     Swap |                    ? | 23.787 ns | 0.5937 ns | 1.7223 ns | 23.275 ns |     - |     - |     - |         - |    1226 B |
|      Dup |                    ? | 12.831 ns | 0.2423 ns | 0.2693 ns | 12.764 ns |     - |     - |     - |         - |    1229 B |
|  **Uint256** |                    **0** | **15.404 ns** | **0.3328 ns** | **0.9549 ns** | **15.232 ns** |     **-** |     **-** |     **-** |         **-** |     **875 B** |
|   Int256 |                    0 | 28.032 ns | 0.5479 ns | 1.0156 ns | 27.713 ns |     - |     - |     - |         - |    1535 B |
|  **Uint256** | **11579(...)39935 [78]** | **16.346 ns** | **0.3488 ns** | **1.0229 ns** | **16.246 ns** |     **-** |     **-** |     **-** |         **-** |     **875 B** |
|   Int256 | 11579(...)39935 [78] | 29.173 ns | 0.6919 ns | 2.0184 ns | 28.504 ns |     - |     - |     - |         - |    1535 B |
|  **Uint256** | **12512(...)57123 [24]** | **14.707 ns** | **0.2842 ns** | **0.2659 ns** | **14.650 ns** |     **-** |     **-** |     **-** |         **-** |     **875 B** |
|   Int256 | 12512(...)57123 [24] | 27.901 ns | 0.5671 ns | 1.1195 ns | 27.590 ns |     - |     - |     - |         - |    1535 B |

### After

|   Method |                    v |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|--------- |--------------------- |----------:|----------:|----------:|------:|------:|------:|----------:|----------:|
|     **Byte** |                    **?** | **10.409 ns** | **0.1982 ns** | **0.1854 ns** |     **-** |     **-** |     **-** |         **-** |     **649 B** |
| PushZero |                    ? |  7.083 ns | 0.1367 ns | 0.1141 ns |     - |     - |     - |         - |     564 B |
|  PushOne |                    ? |  7.940 ns | 0.1621 ns | 0.1592 ns |     - |     - |     - |         - |     566 B |
|     Swap |                    ? |  9.965 ns | 0.2017 ns | 0.1788 ns |     - |     - |     - |         - |     669 B |
|      Dup |                    ? |  9.376 ns | 0.1249 ns | 0.1107 ns |     - |     - |     - |         - |     745 B |
|  **Uint256** |                    **0** | **13.121 ns** | **0.2702 ns** | **0.3787 ns** |     **-** |     **-** |     **-** |         **-** |     **824 B** |
|   Int256 |                    0 | 26.159 ns | 0.5071 ns | 0.4743 ns |     - |     - |     - |         - |    1484 B |
|  **Uint256** | **11579(...)39935 [78]** | **12.831 ns** | **0.1433 ns** | **0.1593 ns** |     **-** |     **-** |     **-** |         **-** |     **824 B** |
|   Int256 | 11579(...)39935 [78] | 26.490 ns | 0.3960 ns | 0.3704 ns |     - |     - |     - |         - |    1484 B |
|  **Uint256** | **12512(...)57123 [24]** | **13.227 ns** | **0.2746 ns** | **0.4809 ns** |     **-** |     **-** |     **-** |         **-** |     **824 B** |
|   Int256 | 12512(...)57123 [24] | 26.414 ns | 0.3757 ns | 0.3331 ns |     - |     - |     - |         - |    1484 B |