@Scooletz 2021-02-22
@tkstanczak I'm leaving it as a draft for a discussion. It's faster and according to my understanding it should be a safe approach comparing to the previous one. We can discuss it and think whether or not this should be pursued. Being given the previous error, I think before merging it, heavier testing is required.