@tkstanczak 2021-02-22
cool thing, I would love to understand where the speed differences are coming from and whether we can remode our Uint256 to be there (maybe we just need to change the alignment of UInt256?)

then we can take UInt256 on / off stack?