@Scooletz 2021-02-22
Sure. Let me keep it open @tkstanczak for now and think about possibly removing the `Word`. The gains are visible. It'd be awesome to have it safe, and ultimately, merged 😃