@Scooletz 2021-02-22
I think I have a reason why `Word` and `Uint256` should be kept separated. `Uint256` is from a separate library. I think that having Nethermind evm implementation based on the external library offsets is not the best thing to do. Also, after tracking `.PopBytes`, which includes treating them as `Keccak` for `LOG` operations, I think this two should be kept separate, even, if they are aligned with the physical layout @tkstanczak