@tkstanczak 2021-02-24
@LukaszRozmej how to best test this? shall we run smoke tests on it?
Would you like to add any extra EVM tests?