@LukaszRozmej 2021-02-24
@Scooletz do you know why previous version worked most of the times and failed only seldom?