@Scooletz 2021-02-24
@LukaszRozmej I didn't get a repro on a failure. My assumption is that with reading a `Span<byte>` to `Word` with all the `Unsafe.Unaligned` was actually missing some alignment, meaning, that it wasn't properly read. This PR changes the way words are handled. Now, the stack is created from words only which are turned into `Span<byte>` whenever needed. This happens for either reporting the value to a tracer or pushing/popping bytes.