@tkstanczak 2021-03-03
@Scooletz can you open a request to merge to a non-master branch on Nethermind?