@Scooletz 2021-03-03
I'm not sure I follow @tkstanczak You mean to retarget this PR or to create a separate one with the same changes? 
Is this related to https://github.com/NethermindEth/nethermind/pull/2834 and you created it on your own?