@Scooletz 2021-03-04
I have green tests from `Ethereum.Blockchain.Block.Test`. This branch requires some good squashing and a bit more clean up. This will be cleaned up tomorrow. 

FYI @tkstanczak