@Scooletz 2021-03-05
This PR is substituted by #2848 , with the reasoning provided in the follow up PR.